Implement Scommit
Ensure commit_msg_filepath is absolute so python will be sure to see the
same file.

Pass a fileobject to commit() so I can test it.

Use exceptions for commit errors.

Handle commit message that's only whitespace.

Add test for committing changes.

Use BufDelete because we know buffer will be deleted and BufHidden isn't
fired when buffer is hidden to be deleted.

Fix autocmd quoting.