Add vim_error_on_fail to API functions
Add it to the ones that are called directly from vimscript. I already
setup the vimscript to handle these errors.

Skipped it on the ones called by mappings that are hooked up in python.
I don't think the error handling will work there.

I also don't want to add it anywhere without setting up the vimscript to
handle the error.