Be more careful about converting paths
Fix 's' on Changelist title causes whole repo to be added. When we hit
the empty line, we'd try to add it which converted to '.' and meant
adding the entire repo.

Now we're very reluctant to return a path from
_get_abs_filepath_from_line unless it seems valid.

Also add a limit of staging 10 files at once. I can remove this when the
code is more solid.