Diff against HAVE by default
Instead of diffing against HEAD (what's on the server), diff against
HAVE. This generally produces more useful diffs (shows your changes) and
is more similar to git (where you don't diff against remote repos).

Also support `:Sdiff % HAVE` instead of an error. svn doesn't accept
HAVE as a revision, but it's nice to be able to use it to be explicit.