Show commit errors in a new buffer
Stop wiping the commit message buffer when it's closed. We cannot open
buffers from BufDelete.

Reuse the commit message buffer so you can get your old messages and we
can use BufWinLeave to open an error buffer when something goes wrong.

We have to be a bit more careful because we always have content in our
message buffer.